REFACTOR: Centralised types
