+++ lib/actions/assistants/list-assistants.ts
@@ -4,7 +4,7 @@ import { requireSession } from "@/lib/actions/require-session";
 import { db } from "@/drizzle/db";
 import { assistant } from "@/drizzle/schema";
 import { eq, desc } from "drizzle-orm";
-import type { AssistantRow } from "./types";
+import type { AssistantRow } from "@/types/assistant-row";
 
 export async function listAssistants(): Promise<AssistantRow[]> {
   const session = await requireSession();
