+++ types/attachment-row.ts
@@ -1,0 +1,10 @@
+export interface AttachmentRow {
+  id: string;
+  messageId: string;
+  userId: string;
+  name: string;
+  mimeType: string;
+  size: number;
+  key: string;
+  createdAt: Date;
+}
