+++ types/chat-with-messages.ts
@@ -1,0 +1,8 @@
+import { ChatRow } from "./chat-row";
+import { MessageRow } from "./message-row";
+import { AttachmentRow } from "./attachment-row";
+
+export interface ChatWithMessages extends ChatRow {
+  messages: MessageRow[];
+  attachments: AttachmentRow[];
+}
