+++ types/mcp-server-config.ts
