+++ types/mcp-server-row.ts
