+++ types/message-row.ts
@@ -1,0 +1,9 @@
+export interface MessageRow {
+  id: string;
+  chatId: string;
+  role: "user" | "assistant" | "system";
+  content: string;
+  parentId: string | null;
+  metadata: string | null;
+  createdAt: Date;
+}
